gh-357  refactored tx service into separate request per file
