+++ Multisig/Data/Services/Safe Transaction Service/SafeTransactionService.swift
@@ -1,0 +1,33 @@
+//
+//  SafTransactionService.swift
+//  Multisig
+//
+//  Created by Moaaz on 5/7/20.
+//  Copyright © 2020 Gnosis Ltd. All rights reserved.
+//
+
+import Foundation
+
+class SafeTransactionService {
+
+    var url: URL
+    private let logger: Logger
+    private let httpClient: JSONHTTPClient
+
+    var jsonDecoder: JSONDecoder {
+        httpClient.jsonDecoder
+    }
+
+    init(url: URL, logger: Logger) {
+        self.url = url
+        self.logger = logger
+        httpClient = JSONHTTPClient(url: url, logger: logger)
+        httpClient.jsonDecoder.dateDecodingStrategy = .backendDateDecodingStrategy
+    }
+
+    @discardableResult
+    func execute<T: JSONRequest>(request: T) throws -> T.ResponseType {
+        try httpClient.execute(request: request)
+    }
+
+}
